SERVER/requirements.txt
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
eventlet==0.33.3
firebase-admin==6.1.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.0
grpcio-status==1.54.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.5
proto-plus==1.22.2
protobuf==4.23.0
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
requests==2.30.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.12
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
Werkzeug==2.3.4
wsproto==1.2.0
zope.event==4.6
zope.interface==6.0
